Move check_expect and check_expect_clear_loc to `inlayhints::chaining`
